Disparse primary key in Spanner
cf. Choose a primary key to prevent hotspots
https://cloud.google.com/spanner/docs/schema-design

+++ configs/config.go
@@ -6,6 +6,6 @@ const (
 	NumUsers = 100000
 )
 
-func RandUserId() int {
-	return rand.Intn(NumUsers/2) + rand.Intn(NumUsers/2) + 1
+func RandUserId() uint32 {
+	return uint32(rand.Intn(NumUsers/2) + rand.Intn(NumUsers/2) + 1)
 }
